Implement ActivityContext for managing activity items and search functionality
