+++ sportsphere-app/App.js
@@ -14,13 +14,15 @@ import ConfirmEditProfileButton from './components/ConfirmEditProfileButton';
 import ReminderScreen from './screens/ReminderScreen';
 import AddReminderButton from './components/AddReminderButton';
 import ChatDetailScreen from './screens/ChatDetailScreen';
+import { ActivityProvider } from './context/ActivityProvider';
 
 export default function App() {
   const Stack = createNativeStackNavigator();
 
   return (
     <View style={styles.container}>
-    <CustomStatusBar statusBgColor={COLORS.primary} barStyle="light-content">
+      <CustomStatusBar statusBgColor={COLORS.primary} barStyle="light-content">
+        <ActivityProvider>
         <NavigationContainer>
         <Stack.Navigator>
           <Stack.Screen
@@ -106,7 +108,8 @@ export default function App() {
             }}
             />
         </Stack.Navigator>
-      </NavigationContainer>
+          </NavigationContainer>
+        </ActivityProvider>
       </CustomStatusBar>
     </View>
   );
